Move Zendesk code to Messenger.php
